feat: add option to choose CPU or GPU test
Add the ability for users to choose between running the matrix multiplication test on the CPU or GPU, using command line arguments. Additionally, the script now accepts an optional argument to specify the matrix size to use in the test. The README.md file has been updated to reflect these changes.